mtda/console/logger: do not maintain line count in rx_buffer
rx_buffer is a dequeue and is configured to not hold more than 1000 lines.
flush() would throw an exception as it was relying on a separate line counter
to determine how many lines needed to be flushed out. Remove this manual
counter and instead use len(self.rx_buffer) where needed.

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>