monitor: add preliminary support for a secondary console
We may need two consoles for some devices: a system control interface
and a regular user console interface. Both will have their own ring
buffers. Add a "monitor send" command to the CLI to test this new
console interface. Tested with a device where a "power on" script
was used to interrupt the boot sequence when the "boot-from-sd"
environment variable is set and then send a custom u-boot command
to boot from the SD card of that device instead of its eMMC. More
commands (and APIs) are probably required to make this new console
interface somewhat useful.

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>