console: support switching between the console and the monitor
The interactive console now supports setups having both a console and a
monitor and switching between them (ctrl-a m). The ZeroMQ publisher now
has a CON(sole) and a MON(itor) topic to stream data from these two
consoles to clients. pause(), resume() and toggle() methods were added
to the console output class. In order to switch to the monitor, the
client will pause the console (data received will be logged into a ring
buffer) and the monitor output resumed.

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>