meta-isar/conf: introduce mtda-bullseye distro
Having per-machine DISTRO_APT_SOURCES or DISTRO_APT_PREFERENCES
was resulting in differences in the isar-bootstrap-target builds
and would cause problems in multi-config builds. Moreover, the
"abstraction" layer provided by the MTDA_EXTRA_APT variables
introduced in c610fef was pointless. A mtda-bullseye distro is
instead introduced (it extends debian-bullseye) and includes
backports we require for NanoPi variants such as the R1.

Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>