support/usb: integrate mtda-usb-functions with mtda itself
Having mtda-usb-functions as a separate service, it could not easily
process configuration of the various functions. It was infact limited
to the path to the block device or file to be used as a shared device.
The mtda.support.usb.Composite class may now be used by drivers to
configure and enable functions that are required. This will be used
to introduce a "monitor" function (as an additional virtual serial
device).

Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>