meta-isar/recipes-python: Build local version
This avoids pulling random versions from github, permitting to package
local changes as well as clearly defined releases. The price is the need
of maintaining a list of top-level files and directories.

As the resulting input for the package build is now a local folder, the
source/format used for the Isar build needs to be switched to default.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>