meta-isar: Add Raspberry Pi 4 model B
This adds a Debian image for the RPi 4, widely based on the previous
work jailhouse-images.

The PPi 4 has a USB gadget port which is shared with the power supply.
Enable this mode by default so that a USB mass storage device can be
emulated via this link, typically with the additional help of a
power/data splitter like https://8086.net/product/usb-pwr-splitter.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>