storage: use tell() to report how many writes were completed
Some decompressors may perform writes in the background and/or
after the final call to write(). Add tell() to the storage driver
interface to report the current position within the storage
instead of calculating it from the some of return values from
write(). This fixes the issue of no progress being shown after
a zst stream was consumed.

Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>