mtda: allow running behind a reverse proxy
This change ensures that all resources are references relative to the
current URI. By that, the whole MTDA web frontend can be served below a
subpath. This is required to support use-cases where multiple MTDA
devices are served by a single nginx reverse proxy in a single web page.
This approach can be used to serve a whole testing network via a single
IP entrypoint and without giving the users IP access to the testing
network itself.

To make the socket.io communication work, we further set the path
component of the current URI when connecting to the ws server.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>