Added algorithm to auto-parse .bmap 2.0 files and write accordingly
In `client.py` a bmap file in the same folder as the image supplied with `mtda-cli storage write image.wic` is searched.
If there is no .bmap  or the parsing of the .bmap file goes wrong,
the image is written just like before. The algorithm was tested with 2 different images with holes.
Each currently supported format (raw, bz, gz, xz and zst) was tested for both of them and successfully flashed to a storage.

Signed-off-by: Manuel Josef Matzinger <manuel.matzinger@siemens.com>