fix(bmap): replace bmap writing algorithm with complete on
The current block-based bmap writing algorithm has correctness bugs
around the handling of partial blocks (overlaps). Fixing these would
require special handling of many corner cases. In addition, the current
algorithm is quite slow as often single blocks need to be written.

This patch completely replaces the algorithm with one that is purely
based on index calculations (no corner cases). This further has the
advantage, that chunks as big as possible can be written, reducing
amount of time spent in python code. By that, we get a significant
performance improvement.

To ensure correctness, we also validate the decompressed image data
against the provided checksums. This adds some overhead, especially on
low-powered devices, but we still believe that correctness is more
important than speed. In total, the performance improvements of the
algorithm itself are significantly bigger than the overhead in the
validation.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>