fix(bmap): reset internal state before writing
Before writing a new image file, we need to reset the internal state of
the bmap writing algorithm. Otherwise subsequent writes do not start
with a clean state. While fixing this, we also shrink the state
by making the number of written blocks a local variable (it can be
trivially computed from the bytes written and the blocksize).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>