feat: auto-detect serial interface of sdwire
The serial of the sdwire interface depends on the physical device it is
attached to. This makes it impossible to pre-configure multiple MTDA
devices for a mass deployment.

This patch auto-detects the serial of an sdwire in case it is not
specified in the configuration. This simplifies the default
configuration and also enables pre-configuration of MTDA devices.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>