fix(mtda-config): reload systemd after changes
The mtda-config service might change systemd dropins. Even in case no
logical change is performed, the timestamps of the files are updated and
by that systemd complains about an inconsistent state. Until we properly
maintain the state and only rewrite configs on actual changes, we simply
always trigger the systemctl daemon-reload.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>