mtda: signal startup completion to systemd
The startup of the mtda service might take quite long on the nanopi-neo
board (~15sec). This is due to the loading of a ton of dependencies,
which cannot really be avoided. To make it possible to depend other
units on a properly running mtda instance, we switch the service type to
signalling and signal when we start the server.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>