fix: avoid races during startup of mtda
Previously, the mtda-config.service and mtda.service were not
synchronized, leading to a restart of the mtda service during the
startup of it. The first part of this patch adds a strict (time)
synchronization between the two.

As the mtda-systemd-helper first needs to finalize the systemd
configuration before the next job can be scheduled, we also need
to switch the service to oneshot (wait for completion before
becoming active). This however creates a deadlock in case we
synchronously wait for the completion of the mtda-config service.
To fix this, we trigger the "try-restart" in async mode (--no-block).
By that, the mtda-config service will complete first and afterwards
the mtda service will be started. This also avoids the double
restart on startup, as the restart is simply merged with the
following startup by systemd.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>