fix(client): clamp heartbeat to timeout value
When using RPC timeouts shorter than the heartbeat value, the timeout
of the RPC call was not properly working, as the call only returned
after the heartbeat duration. In normal operation, this was not an
issue, as the heartbeat was significantly shorter than the timeout, and
also timeouts only happened due to function invocation timeouts, but not
due to connectivity losses.

When working with one-shot RPCs (i.e. create a client to only call one
rpc), this led to huge delays and an accumulation of half-open TCP
connections. By clamping (limiting) the heartbeat duration to the
timeout, also short timeouts are properly supported now.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>