perf: make make_printable more efficient
This cuts the initial MTDA startup time in half by avoiding the creation
of a huge replacement table for string ASCII encoding. Instead, we use
the encode infrastructure directly (which is also more efficient on the
replacement itself).

Note, that this is technically not a 1:1 replacement, as non-printable
characters are replaced by a '?' instead of a '.'. However, as this is
anyways only used in the debug output, we believe it is a reasonable
tradeoff.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>