feat: compress image on-the-fly
When copying an uncompressed image, we lightly compress the image on the
fly with zstd on transfer. By that, subsequent zeros (holes) are
automatically collapsed and not transferred, which can give a huge
speedup on uncompressed images. Another advantage is, that zstandard can
be decompressed with a very low overhead (e.g. compared to gz or xz) and
is better suited for low-memory devices.

closes: #20

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>