fix(pytest): do not fail if the default TEST_IMAGE cannot be found
Target in our pytest module expects the image specified in the TEST_IMAGE
environment variable (or its default value) to exist. No longer fail if
an image is not specified in the environment and if the default one could
not be found: this may be intented.

Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>
(cherry picked from commit f9a0facb7afc4e513dbfe16d12954d55865aa319)