fix(www): do not call storage-flush endpoint
We have different queues for the data and control planes: the
javascript code may have sent the entire file to the websocket
but the browser will have it enqueued into its own queue. With
calls we make being asynchronous, remove the unnecessary call
to the storage-flush endpoint: we will be notified when the
agent has finished processing the queue (with a CORRUPTED or
INITIALIZED event).

Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>