fix(mtda-service): create home dir of user mtda
In the ISAR integration, we change the mtda user to be used for
interactive logins. However, this requires a homedir, as usermod does
not create a homedir for a user that did not have one before (which is
also correct, as the home bootstrapping only happens on adduser).

As the ISAR user creation logic is executed after the rootfs install,
the postinst script of mtda-service does not know about the mtda user
and creates no homedir. Then, when ISAR changes the user, it was created
without a home, hence the changed user also not gets any home. By that,
the mtda user remains without a home, which leads to various warnings on
login.

We fix this by always creating the mtda user with a homedir.

Fixes: 5bf7a48 ("fix(www): switch from flask to tornado")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>