feat(rpi4): add target to support EFI booting
As a preparation to support OTA updates via the swupdate pattern
implemented in isar-cip-core, we enable EFI booting of the Raspberry Pi
4 B. The implementation boots the EDK2 variant for the RPi as a second
stage, loads and patches the DTB and then performs standard EFI booting
(using systemd-boot).

Due to limitations in the used EDK2 implementation, currently only
booting from USB is supported. Booting from the SD card works as well,
but is horribly slow as the controller runs the card in the "high-speed"
mode during boot.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>