feat(rpi): enable UART2 interface
As we use an upstream kernel with its device tree, the RPi overlays do
not apply. Instead, we provide a custom overlay to enable the UART2 on
Linux 6.12. While we are at it, we also remove the not-needed dwc
overlay for the USB gadget mode (and custom kms), as this is
automatically enabled upstream.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>