feat(rpi): switch to U-Boot as EFI provider
The EDK2 implementation is far from being usable. It does not support to
pass a device tree. Instead, the device tree is converted to ACPI tables,
but this conversion is not feature complete. Even simple things like
GPIO access is impossible to achieve. We now switch to U-Boot as EFI
provider as this has proper support for device trees. We further use the
the DT from Linux and by that have to rewrite some of the overlays.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>