feat: add a/b updatable image
We add an image variant that can be updated with the A/B rootfs update
scheme from CIP. As these images are blueprints for mass deployments, we
remove the fixed mtda hostname, as the hostnames need to be assigned by
an DHCP server. We further do not add any swupdate backend integration
yet, as this is use-case specific.

To avoid having a dependency to CIP in the mtda-isar layer, we add these
parts in a dedicated mtda-isar-extra layer.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>