fix(image): correctly handle range end with overlap
There was a subtle blug in our implementation of the bmap algorithm: When
ending a range but still carying an overlap (a not fully written block),
the next overlap is computed was computed incorrectly. By that, the
writing position shifted by [0-blocksize) bytes which lead to
incorrectly written data blocks. This was detected by a checksum
mismatch.

We fix this by further simplifying the algorithm: Instead of computing
the overlap based on a delta, we simply compute it by taking
"writtenBytes % blocksize". We did this anyway in case the current input
was fully processed, now we also do it within the block writing logic.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>