client: check if image is fully written
In case the storage write is interrupted (e.g. because of a checksum
mismatch or another internal error), we need to inform the other side
about the issue.

As the current architecture does not provide an easy mechanism to inform
about this, we just check if the written bytes are equal to the expected
ones (in case we know about the full size.

Closes: #476

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>