fix(www): make get() async to handle possibly blocking calls to MTDA
Tornado uses co-routines and we may not block in the get() method of
our handlers. Delegate calls to MTDA to a thread pool via a common
blocking_call() method. The number of worker threads may be adjusted
in the www configuration fragment.

These changes fix some sporadic hangs observed with the /power-toggle
endpoint (but could probably be reproduced/seen with any other endpoints
really).

Refactor StorageToggleHandler by adding storage_toggle() to the exposed
Pyro APIs: this allows that endpoint to follow the same structure as all
others (single remote call).

Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>