fix(www): make pyodide pass line feeds to xterm.js
The pyodide stdout/stderr handlers may be configured in batched,
raw, or write modes. In batched mode (which is the default mode),
pyodide would call our handler for each line or flush but line
feeds do not seem to be included when they should. Use the raw
mode instead to be called each time a character needs to be
printed to the terminal.

Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>