feat(kas): add abrootfs support on trixie
On trixie, the SWUs need to be signed for technical reasons. We add
support for this, by signing with the debian snakeoil keys. If needed,
custom keys can be used as well. The update path from bookworm to trixie
still works, as our bookworm swupdate also accepts signed update files
and just ignores the signature.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>