fix(abrootfs): assign stable uid / gid for mtda
We currently do not assign a stable uid / gid for the mtda user / group.
By that, each install might get a different one. In case of abrootfs, we
need stable assignments, as the ids need to be stable across multiple
rootfs creations, as persisted data (like the homedir and mtda managed
files) otherwise cannot be accessed after an update.

Fixes: 04ad2d5 ("feat: add a/b updatable image")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>