feat(mtda): add command to apply firmware update
This is a first and rough attempt to allow updating a ab-rootfs
installed MTDA image from the mtda-cli. We implement this around the
image streaming API, as it is semantically quite similar. However, due
to the MTDA internals at some locations we need to cut corners. Further,
there is currently no mechanism to confirm an update after the reboot
(will be added, once the other points are clarified).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>