fix(storage): break on storage errors during transmission
Errors that happened while writing to the storage remained unnoticed to
the client in case they happened during data transmission: The error was
detected on the receiving side, which closed the ZMQ socket. This
resulted in an infinite block on the sending side.

We fix it by using the nonblocking send API. In case the sending fails
because the queue is full, we check the storage status for write errors.
As a full queue is not per-se an error (esp. on slow networks), we
implement a simple exponential backoff algorithm and also only check the
status if we can't send data for too long (to avoid frequent RPC calls).

Xref: #501
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>