chore(storage): only rely on writing events for progress
We already get the progress via the event channel and also call the
progress reporting callback there. By that, we no longer explicitly need
to check the storage status for progress (but we need to do it for
correctness / status).

We now drop the callback invocations along with the now empty progress
method.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>