fix(storage): issue write event prior to termination event
In a617348, a final storage write event was added to ensure the progress
bar completes instead of stopping below 100%. This however broke the
state machine of the web UI, as it uses the INITIALIZED / CORRUPTED
event to inform the user about the termination of the storage write.

We fix this by moving the final write event into the writer itself and
prior to the termination storage event. A nice side effect is a
simplified code with less cross-module calls.

Fixes: a617348 ("perf(cli): use event channel for "storage write")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>