refactor(storage): avoid call to storage_status on flush
We only need to report if the storage write operation was successful or
not. As we already get this information from the storage_flush method,
we don't need to call storage_status. By that we can't report the
exact position at which the storage write failed, but if this is of any
interest in the future, we still can get it from the final storage write
event.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>