refactor(rpi4b-efi): switch target over to rpi-arm64-v8-efi
We now have a identical copy of our rpi4b-efi target in upstream ISAR.
The WIC file slightly differs, but this does not suffice to have a full
copy of the target in the mtda layer. By that, we replace our rpi4b-efi
target with the rpi-arm64-v8-efi target from upstream ISAR.

I tested the new image on a rpi4b.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>