feat(storage): make progress reporting more linear
When writing with a bmap, the storage write operation itself is not
linear in time, as seeks are much quicker (basically free) compared to
actual storage writes. On sparse images this makes a huge difference.
Further, reporting the progress based on the data that is consumed by
the input stream is very imprecise, as this is potentially compressed
data. By that, the progress bar was jumpy and also the reported speed
basically meaningless.

We change this in an ABI conserving way by splitting the writes into the
actual written bytes and the seeks: The old storage status API just gets
the combined value (so that downstream users can compare for equality
between the image size and the written bytes), but the event based
progress reporting operates on the writes only. If we further have
information about the total number of mapped bytes (we have them when
writing with a bmap), we give precise progress information.

With these changes, the progress bar is quite linear and the speed
reflects how fast we can write to disk.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>