fix(storage): validate checksum of last range as well
We currently only validate a range once it gets incremented. However, in
case of the last range, we do not increment, as this would lead to an
out-of-bounds error in the bmap dict.

We fix this by splitting the range-bump part and the validate range
part. We now validate directly after finishing a range and bump on the
next iteration.

Fixes: 77e6715 ("fix(bmap): replace bmap writing algorithm with ...")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>