fix(storage): validate last block range on partial blocks
In case the total image size is not a multiple of the blocksize, the
last written block is underfull. By that, we don't validate the range.
In the current implementation this remains unnoticed, as the data
stream is closed with the correct number of written bytes.
Note, that the image is fully written, but just the check for
correctness is missing.

We fix this by checking if the total number of written (or skipped)
bytes is equal to the image size and validate the range in this case as
well.

Fixes: 77e6715 ("fix(bmap): replace bmap writing algorithm with ...")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>