feat(storage): implement commit/rollback for qemu
QEMU images may be created as CoW image on top of a raw image using
qemu-img create. If the qemu storage is configured to have a cow
image, a qcow2 image will be created and used as a drive. "commit"
is implemented using "qemu-img commit" and will cause changes written
to the qcow2 image to be written to the base raw image. For rollback,
we simply create a new qcow2 image.

Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>