fix: storage_flush check thread exists before joining
Some issue are observed during storage write, happens when
the storage_flush is called by client. At server thread join is tried
when flush is called in writer, since the stop has already been called
thread for the instance is already reinitialized to none, Due to this
the at client we observe errors stating calling join on None type.

Tested-by: Gourav Singh <gourav.singh.ext@siemens.com>
Signed-off-by: Shivaschandra KL <shivaschandra.k-l@siemens.com>